+++ src/main/java/com/preetu/backend/service/UserService.java
@@ -2,11 +2,15 @@ package com.preetu.backend.service;
 
 import java.util.List;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.preetu.backend.dto.UserRequest;
 import com.preetu.backend.dto.UserResponse;
 import com.preetu.backend.entity.Users;
+import com.preetu.backend.exception.ConflictException;
+import com.preetu.backend.exception.ResourceNotFoundException;
 import com.preetu.backend.repository.UserRepository;
 
 @Service
@@ -18,8 +22,15 @@ public class UserService {
 		this.userRepository = userRepository;
 	}
 
+	// CREATE USER
 	public UserResponse createUser(UserRequest request) {
+
+		if (userRepository.existsByEmail(request.getEmail())) {
+			throw new ConflictException("User with email already exists+ " + request.getEmail());
+		}
+
 		Users users = new Users();
+
 		users.setName(request.getName());
 		users.setEmail(request.getEmail());
 		users.setPhone(request.getPhone());
@@ -27,37 +38,52 @@ public class UserService {
 		Users savedUsers = userRepository.save(users);
 
 		return toResponse(savedUsers);
+	}
+
+	// GET ALL USERS
+	public Page<UserResponse> getAllUsers(Pageable pageable) {
+
+		return userRepository.findAll(pageable).map(this::toResponse);
 
 	}
 
-	public List<UserResponse> getAllUsers() {
-		return userRepository.findAll().stream().map(this::toResponse).toList();
-	}
-
+	// GET USER BY ID
 	public UserResponse getUserById(Long id) {
 
-		Users users = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found."));
+		Users users = userRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
+
 		return toResponse(users);
 	}
 
+	// UPDATE USER
 	public UserResponse updateUserById(Long id, UserRequest updatedUser) {
-		Users existingUser = userRepository.findById(id).orElseThrow(() -> new RuntimeException("Invalid ID."));
+
+		Users existingUser = userRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
 
 		existingUser.setName(updatedUser.getName());
 		existingUser.setEmail(updatedUser.getEmail());
 		existingUser.setPhone(updatedUser.getPhone());
-		
+
 		Users updatedUsers = userRepository.save(existingUser);
 
 		return toResponse(updatedUsers);
-
 	}
 
+	// DELETE USER BY ID
 	public void deleteUserById(Long id) {
+
+		if (!userRepository.existsById(id)) {
+			throw new ResourceNotFoundException("User not found with id: " + id);
+		}
+
 		userRepository.deleteById(id);
 	}
 
-	private UserResponse toResponse(Users users) {
-		return new UserResponse(users.getId(), users.getName(), users.getEmail(), users.getPhone());
+	// ENTITY → RESPONSE DTO
+	private UserResponse toResponse(Users user) {
+
+		return new UserResponse(user.getId(), user.getName(), user.getEmail(), user.getPhone());
 	}
 }
